frontend/requirements.txt
streamlit>=1.28.0
plotly>=5.17.0
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
requests>=2.31.0
supabase>=2.0.3
PyJWT>=2.8.0,<3.0.0
langchain>=0.1.16
langchain-core>=0.1.48
langchain-community>=0.0.38
langchain-google-genai>=0.0.6
google-generativeai>=0.4.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
jinja2>=3.1.4
python-multipart>=0.0.9
pillow>=10.0.0
pytesseract>=0.3.10
easyocr>=1.7.1